Support non-indexed arrays, improved performance
Support to edit non-indexed arrays. The array will be stored as ServerPackages, ServerPackages{1}, ServerPackages{2} etc. This results in a small difference in setting values. If an array value is set that doesn't exist an non-indexed value is added, indexed array keys can be set by including the index in the key.

The performance has been improved by storing keys in a hash rather than an array. Support for non-indexed arrays was a requirement for this. The order of keys is saved in a special array just like the order of headings.